Through sub_80417B8

+++ include/daycare.h
@@ -3,7 +3,7 @@
 
 u8 Daycare_CountPokemon(struct BoxPokemon *);
 void sub_8041324(struct BoxPokemon *, struct DayCareMailEtc *);
-void sub_8041790(int i);
+void sub_8041790(u16 i);
 u16 sub_8041870(u16);
 void sub_8041940(void);
 void sub_8041950(void);
